requirements.txt
flask
pymongo
bcrypt
pyjwt
